scripts/requirements.txt
requests==2.32.4
python-dotenv==1.1.1
cloudscraper==1.2.71
pandas>=2.0.0
lxml>=5.0.0